requirements.txt
flask==2.3.3
Werkzeug==2.3.7
Pillow==10.0.0
torch==2.0.1
torchvision==0.15.2
transformers==4.32.1
matplotlib==3.7.2
numpy==1.25.2
